Session handling and logout
Sessions are handled through express-session middleware. They default to
using an in-memory store so they are wiped whenever the node process is
restarted, but they can trivially be configured to rely on redis
instead. The default in-memory option is only suitable for development.

The ability to log out of a current session is added for /logout.